[FLINK-16661] Consolidating code in ApplicationClusterEntrypoints
